Fixes Load Problem
A problem occurred when a file was loaded before its dependencies, it was fixed
by adding the files with problem to a list, load the other files and try to
reload them.

To avoid an infinite loop, the application retries to load the list of files 3
times and skips the remaining unloaded files after this.